Wifi: Enable supplicant's auto reconnect ability if auto join is disabled
Auto Join capability in the framework shall take the control of WiFi
network management and thus does not allow the supplicant to auto
reconnect to the configured profiles. However,to ensure automatic
connection to the networks happen when auto join functionality is
disabled in the framework,supplicant needs to be configured with auto
reconnect capability.
This commit ensures the same.

Change-Id: I08119f459b9c4cc7a022f6c3e0b881b1542c671c
CRs-Fixed: 759411