wifi: Address the improper scan behavior on PNO failure
This commit takes care of the following behavior on a PNO fail.

Trigger a periodic scan on a PNO failure when p2p session is active
and there are no configured profiles.

There is no point to continue the periodic scan triggered on a PNO fail
when p2p session is active.Thus, stall further attempts after a p2p
disconnections.

Change-Id: I7844dca7fe9dd45b99e71ef04beb19ea2ae5343e
CRs-Fixed: 759618