wifi: Provide fix for no connect status for a network with WPA-PSK followed by WPS
WifiConfigStore maintains the state of each configured Wifi profile.
These profiles are maintained to create a one-to-one mapping between
the profile and the IP configuration. The same list is also used
to display the connected profiles.

While creating list of valid network configurations, the profiles
with same SSID but different security and authentication credentials
are considered duplicate. Thus the profile created due to WPS
connection attempt is not added to the list maintained by WifiConfigStore.
This results in the framework not displaying the new network as connected,
as the supplicant shall give the connect indication on the new network id
which shall not get added. Thus, overwrite the previous duplicate-disabled
configuration to fix this issue.

CRs-Fixed: 760505
Change-Id: Id1dc26733b9b5ead402e3a2ae4aed3313568c727