WiFi: If auto join is disabled then issue scan with connection
If Auto join is disabled then supplicant in disconnected state. Supplicant
in disconnected state shall attempt to connect to the configured profiles
based on the scan requests triggered by the framework. This requires the
framework to issue scan with connection setup. So issue the scan with
connection setup when station is disconnected.

Change-Id: Id74a12b0b30baee68f99e7c25d3e435c537234c8
CRs-Fixed: 764614