wifi:Provide UI with SIM info for EAP-SIM authentication
User Interface requires the SIM information (viz. number / type
of SIM's supported on the device), using which the respective
authentication option is provisioned to the user. This commit
queries the same by invoking a command exposed by the
wpa_supplicant. It is the wpa_supplicant that gets this
information

CRs-Fixed: 775942
Change-Id: I93c6a5450c806d8fade0103ef96eca986c925eba