WiFi : Indicate Start/Stop of DHCP phase to the driver for BT coexistence
BT when coexists with WLAN shall interfere with the critical DHCP phase of
station / p2p connection acquiring the WLAN slot time, thereby resulting in
to the failure of the critical DHCP operations. This requires an indication
of DHCP start / stop to the firmware so that the WLAN module is given
preference over BT. While this is handled for the station , p2p connection
missed indicating the same and this commit fills the gap.

Change-Id: Iea712d793056fc906a3444ec2d04e9bc9385bda5
CRs-Fixed: 766982