wifi:Truncate ssid_utf8 to BUF_SIZE
Current code does a strncpy from dest buffer of size 2048
bytes to ssid_utf8 buffer of size 256 bytes , while the data
in the dest buffer shall not exceed 252 bytes. To solve this
the length for strncpy is limited to BUF_SIZE. Also,the option
for dynamic memory allocation /free of the buffers does not
result to be foolproof.Hence, revert these changes.

Change-Id: I3ec310b64412069372ea769a6150935a4d86b0e9
CRs-Fixed: 778362