WiFi: Provide fix for connection failure with saved profile
Commit Id1dc26733b9b5ead402e3a2ae4aed3313568c727 ensures to overwrite a
duplicate network profile with the newly found network configuration
provided its status is CURRENT , which indicates that this is the
configuration with which the device is connected. While this addresses the
scenario of - no connect status for a network with WPA-PSK followed by WPS, the
scenario of two such entries present in the supplicant 's configuration shall
result in a connection failure on a reboot as the later valid network shall not
overwrite prior invalid network, for the later network profile is not in the
CURRENT state, during the load process.
Hence, enhance the condition to also consider the status as ENABLED while
overwriting such duplicate configurations.

Change-Id: Iff771a9a045d3e807ad0935cc3a6d3e509bf2a0c
CRs-Fixed: 779407