WiFi: If DHCP failed when auto join disabled, issue a RECONNECT.
If DHCP failed when auto join disabled, DISCONNECT is issued to
supplicant, which makes the disconnect flag enabled, thus SCAN
commands to supplicant will not trigger connection afterwards.

In this case, immediately issue a RECONNECT command to supplicant
to clear the disconnect flag, and allow supplicant to connect to
any save network profiles on receiving SCAN commands.

Change-Id: I0a308cc39b4b0ac21e83aa3a3525d4ab2f5dcaa3
CRs-Fixed: 780306