wifi: Do not update conf file with the default sim_num value
When Wi-Fi is turned On with saved profile, the framework
will over write the existing sim_num value with the default
value. Due to which, the supplicant will always try to
authenticate with the default sim_num value (i.e., 1).
To mitigate this, update the conf file only if sim_num has
non zero values

Change-Id: I2c17944d44f5937974593c91ef27ed0dc39ece38
CRs-Fixed: 780903