WiFi: Stop monitoring on socket if interface is down
When Wifi is turned off, WifiStateMachine issues terminate command to the
supplicant and waits for the event CTRL-EVENT-TERMINATING. On receiving this
TERMINTING event, WifiMonitor stops further monitoring of events. In some
instances, WifiMonitor fails to receive the event CTRL-EVENT-TERMINATING and
subsequently when Wifi framework times out it kills the supplicant. But this
recovery logic has an issue where even though the supplicant is killed, the
monitor thread is not stopped from listening the events and WifiMonitor
continues looking for the events. Since at the time of killing the supplicant
monitor_conn is initialized to NULL, internally fabricated
WPA_EVENT_TERMINATING event is issued by the framework. This again leads to
the deinitialization of the wlan0 / p2p0 interfaces when the STA / P2P
interfaces gets initialized (during Wifi turn ON). Framework is not designed
to concurrently handle the control and monitor sockets. To solve this issue,
force stop the monitor state machine when Wifi framework kills the supplicant
explicitly

Change-Id: I4045ef226dbd2e627613e0524de45bb723127038
CRs-Fixed: 788582