Wifi:Do not enable networks in DisconnectedState if autoconnect is disabled
WiFiStateMachine enables saved profiles when enters into DisconnectedState.
Due to this device connects to saved profile, even though autoconnect is
disabled. This commit ensures not to enable saved profiles, if autoconnect
is disabled.

This commit also ensures that the framework logic does not depend on
existActiveNetwork check while deciding for AutoConnect. This is in agreement
that the User Interface shall take the decision to set the autoConnectPolicy
through the WIFI_AUTO_CONNECT_TYPE flag based on the existing Active Networks.

Change-Id: I78de7b92f9e764acba03a135759b28b02060ca79
CRs-Fixed: 789010