Clear link debounce on connect event in connected state for no Auto Join
WifiStateMachine enables link debouncing logic (deferring the disconnection
for 7 seconds), if a disconnection event is received in connected state to
avoid spurious disconnection events, thus providing an option to reconnect
to the network during this period. Framework shall attempt to connect to
the network during this period and shall not further proceed to disconnect
if a connection succeeds. While this logic holds good when Auto Join
functionality is enabled, this commit shall address the case when Auto Join
functionality is disabled as in this scenario supplicant is given the
control to reconnect to other networks and thus a connect indication shall
be given to the framework when the WiFi is still in the L2ConnectedState.
Thus, the link debouncing is cleared when a connect indication is obtained
in the L2ConnectedState for Auto Join Disabled, link debouncing
functionality enabled and debouncing logic already in progress scenario.

Change-Id: I24b57d2059b82f2f1b9636068c114fa4fb35680b
CRs-Fixed: 819583