wifi: issue RECONNECT when enter connect mode if autojoin disabled
When "Scan always available" is enabled, supplicant will keep running,
enable / disable Wi-Fi makes framework switch between scan only mode
and connect mode.

When framework enter scan mode, it will issue 'DISCONNECT' command to
supplicant, which makes supplicant transition to disconnected state.
When framework back to connect mode, with autojoin disabled, supplicant
is supposed to trigger the connection based on scan results, but since
it is in disconnected state, no connection will be triggered as expected.

This change is to issue a 'RECONNECT' command to supplicant when enter
connect mode if autojoin is disabled, which allows supplicant to
trigger the connection.

Change-Id: I063739f7c8241ed6a11e94dd7bd85182d6e5dc90
CRs-Fixed: 887616