3-port: Stop Supplicant in case of wifi off state
Existing implementation in 3-port concurrency disables the
available interfaces when wifi is turned off. This is leading
to failure in starting supplicant next time when wifi is turned
on as we receive supplicant status as stopped even when we start
the supplicant service. The change results in stopping the
supplicant when wifi is turned off.

CRs-Fixed: 949970
Change-Id: I7c189d82456f2056f505a237ff0a3791eaaa6083