Update default route MAC address on CMD_NETWORK_STATUS  indication
WiFiStateMachine is updating the default route mac address while entering
the ConnectedState, resulting in the transmission of the TCP Echo frame.
At this point of time the default interface at the ConnectivityService is
not the Wi-Fi interface and thus this frame is transmitted through the
Modem as the cellular is still the default route at this point. This
results in the call getting dropped on entering the Wi-Fi zone from WCDMA.

Thus, address this issue by updating the default route after getting the
CMD_NETWORK_STATUS indication, which is further issued after the captive
portal check, at which point the default interface shall be Wi-Fi.

Change-Id: I82ac29c576932bea710554e92874f8449682aaba
CRs-Fixed: 954064