Wifi: Clear lastConnectAttemptTimestamp variable on a Wifi turn off
Framework shall defer the SCAN request if issued with in 10 seconds after
the last connect attempt. This logic is driven by
lastConnectAttemptTimestamp, which holds the last connection attempt's time
stamp. For a scenario where a WiFi is turned off during the connection
attempt, this variable ended up retaining the last connection time stamp
and thus, the first scan after the WiFi turn on gets deferred, if the
attempt happens with in 10 seconds, resulting in a delay to display the
scan results after the Turn ON.

Hence clear lastConnectAttemptTimestamp variable, when wifi is turned off.

Change-Id: Ib3f306ba38debd95e047cd2ef2fdadf8d9fe3367
CRs-Fixed: 955289