Wifi: handle supplicant state change during linkdebouncing
In case of linkDeboucing, framework will defer the disconnection
for 7 seconds, and allow reconnect thru the framework roaming.
In the RoamingState, supplicant state change is handled without
updating SSID in mWifiInfo.

This change is to return NOT_HANDLED in RoamingState, which allow
ConnectModeState to update the SSID of mWifiInfo.

Change-Id: I110f7916f2203b1cb6fb740b88665b79ee7e3872
CRs-Fixed: 956442