wificonfig: store the scan cache for saved wifi configuration
Each saved profile has an entry in networkHistory.txt file placed
at /data/misc/wifi location whose information is later used to
autojoin when wifi is turned on. The logic to read this file skips
the logic to store the scan details of that particular wifi configuration.
Hence the parsing logic is modified to store the scan cache for
each saved wifi configuration .

CRs-Fixed: 958484
Change-Id: I6627cda3810b56b766664992676cdf1441c77fe9