Wifi: Trigger Selective Scan based on saved profile
This is a new requirement intended to decrease the reconnection
time with saved profiles when Wifi is turn On. The approach
here is to prepare a channel list based on the saved profile
and issue this selective frequency scans for the first two scans.
Moreover inter scan period is reduced to 2 seconds which is a
configurable parameter.
Third scan onwards it issues a full scan.

Change-Id: I72edd47aed7741745d8ebf663e53192be3405e49
CRs-Fixed: 909098