wifi: Use isGbkString() to check if SSID is GBK encoded
Wifi gbk2utf module used weak logic to check if SSID is GBK encoded,
it has chance that SSID wrongly converted and truncated, which in
turn caused framework crash due to unhandled Exception.

This change is to add isGkbString() to check if SSID is GBK encoded,
which makes sure only SSID with GBK encoded will be converted.

Change-Id: I6cb608bfb6f0930d89a58d66114eabbf856acc9d
CRs-Fixed: 978452