Issue scan irrespective of screen on/off cond. upon recv. Network Disconnect event
Existing implementation issues scan for the connected AP
in screen on condition only when network discoonection event
is received in connected state.
In case of VOWIFI when screen is off and we have two bssids
having same ssid name and then one of the bssid is turned off
then scan is not attempted as the screen is off. The check for
screen on/off condition is removed to enable AP to AP Handover.

CRs-Fixed: 1082334

Change-Id: Ic32c9c69fa267ef9441d96a375651f6188295463