HS2Rel1: Trim ScanDetailCache entry to avoid re-selection of same HS AP
When 2 or more HS2Rel1 supported APs are seen which are having
matching homeSP information but different SSID, they will be mapped
to one ScanDetailCache. This will lead to selection of first entry
for setting SSID parameter even though that AP is not seen for long
time. This will cause problem in connection due to SSID mismatch.

Before setting SSID of passpoint network, remove the ScanDetail
entries for the APs which are not seen for sometime [currently 30 sec].
This will aovid setting incorrect SSID if that AP is not nearby.

CRs-Fixed: 1106602
Change-Id: I6dff81113dd45bd4c9cdc2d7d3eae79ee6c96cbd