StaSapConcurrency: Set flag mStaAndAPConcurrency post object instantiation
WifiStateMachine can ran into a scenario where it was trying to access
an object mSoftApStateMachine which is yet to be initialized in other
thread. A flag mStaAndAPConcurrency was used to signify if
mSoftApStateMachine initialization has already completed or not.
Currently mStaAndAPConcurrency is being set before creating the object.
In a situation where mStaAndAPConcurrency is set and mSoftApStateMachine
is being initialized, if another thread tries to access this object, it
will throw Null pointer exception.

Fix this by setting flag mStaAndAPConcurrency, post mSoftApStateMachine
object initialization.

While there, correct minor indentation.

CRs-Fixed: 2021899
Change-Id: Ib2f3cd1af1bae2a6b0564aa7a12fb322d81f1b87