wifi: Trigger stopwifi during the startwifi.
Commit I3e4dd6cd06b1cdaed69ae044cf619a1a431fe7e7 removed the stopWiFI
done during the IWifi registration notification to address a failure
to turn On Wifi during the bootup. This commit introduced a regression
for a use case where the Wi-Fi is expected to work on a system server
restart. This is because of the stopWiFi not getting invoked before the
startWifi. Address this issue having the stopWiFi triggered before the
startWiFi. This commit also retries the startWiFi attempts .

CRs-Fixed: 2149106
Change-Id: Ie904217718552d888b48374339a9289b59822c16