wifi: Add extra Gbk Ssid to hidden networks for scan
When non-hidden AP uses Chinese Ssid with Gbk encoding, the AP is
present in scan results and framework can auto connect via wifigbk.
But if AP changed to hidden, AP will not be present in scan results,
wifigbk can not help, so framework can not auto connect.

When framework sends scan request, add an extra Gbk Ssid to hidden
networks if the ssid is non-ascii (maybe Gbk).

Change-Id: I0d62b18725e2b9253890e23b2b84a27dc539c16e
CRs-Fixed: 2184025