wifi: Do not modify message in QcStaDisablingState for CMD_SET_AP.
In transient state QcStaDisabling, when a request to toggle softap
(CMD_SET_AP) is received, it modifies the msg.arg2 to 1 which would
avoid updating wifi STA last state. For standalone STA/SAP case,
when wifi and softap is toggled alternatively in quick succession
where WifiController is in QcStaDisablingState, a request to start
softap (CMD_SET_AP) would not modify last STA state to DISABLED.
Hence a subsequent softap off would try to start STA again.

Since this transient state is shared in both standalone and STA +
SAP concurrency, we need not update the msg.arg2. softap start
caller is already taking care of this.

CRs-Fixed: 2331418
Change-Id: I95bd4b2f8d7412af2523ecc9feffcbfd844848e0