wifi: Fix connecting endlessly issue if input wrong password
When 4-way handshake is ongoing, mStateIsFourway from vendor hal
is set to true. When handshake fail, onDisconnected() of native
hal is invoked, so mStateIsFourway of native hal is checked. This
is not expected.

The solution is to move mStateIsFourway to a common place, so that
native & vendor hal can share same instance.

Change-Id: Idcf569c83554eef9ad462b8977c4a0b2850df36c
CRs-Fixed: 2532959