HalDeviceManager: Allow NAN to bring down P2P iface
Assign P2P & NAN ifaces the same priority. So, they're able to bring
each other down, but nothing else.

Also,
a) Fixed a bug in the proposal creation exposed by this change. The
list of ifaces to be deleted for a proposal needs to be accumulated
across all iface types. This bug was causing a failure in
testInterfaceCreationFlowTestChipV2 because HalDevMgr was choosing a
proposal which would delete 2 ifaces (NAN + 2nd STA for an AP)
instead of choosing the best proposal which was to delete 1 iface (2nd
STA for an AP).
b) Few log message cleanups to make it more clearer (added a
toString, Arrays.toSring, etc)
c) Fixed all the unit tests to account for the new priority.

Bug: 153842431
Test: atest com.android.server.wifi
Change-Id: I06ca1f18347a8cbe8c495f4540365ed9c34ad009
(cherry picked from commit f85ddd3f8f0f2cba9bf1194983424e98088c5e11)